Move login state to App
